+++ libs/eavmlib/src/spi.erl
@@ -42,17 +42,19 @@
 
 -export([open/1, close/1, read_at/4, write_at/5, write/3, write_read/3]).
 
--type spi_peripheral() :: hspi | vspi.
+-type peripheral() :: hspi | vspi.
 -type bus_config() :: [
-    {miso_io_num, non_neg_integer()}
-    | {mosi_io_num, non_neg_integer()}
-    | {sclk_io_num, non_neg_integer()}
-    | {spi_peripheral, spi_peripheral()}
+    {poci, non_neg_integer()}
+    | {pico, non_neg_integer()}
+    | {miso, non_neg_integer()}
+    | {mosi, non_neg_integer()}
+    | {sclk, non_neg_integer()}
+    | {peripheral, peripheral()}
 ].
 -type device_config() :: [
     {clock_speed_hz, non_neg_integer()}
     | {mode, 0..3}
-    | {spi_cs_io_num, non_neg_integer()}
+    | {cs, non_neg_integer()}
     | {address_len_bits, 0..64}
     | {command_len_bits, 0..16}
 ].
@@ -93,10 +95,10 @@
 %%
 %% <table>
 %%   <tr> <th>Key</th> <th>Type</th> <th>Default</th> <th>Description</th> </tr>
-%%   <tr> <td>`miso_io_num'</td> <td>`non_neg_integer()'</td> <td>-</td> <td>MISO pin number</td></tr>
-%%   <tr> <td>`mosi_io_num'</td> <td>`non_neg_integer()'</td> <td>-</td> <td>MOSI pin number</td></tr>
-%%   <tr> <td>`sclk_io_num'</td> <td>`non_neg_integer()'</td> <td>-</td> <td>SCLK pin number</td></tr>
-%%   <tr> <td>`spi_peripheral'</td> <td>`hspi | vspi'</td> <td>`hspi'</td> <td>SPI Peripheral (ESP32 only)</td></tr>
+%%   <tr> <td>`miso'</td> <td>`non_neg_integer()'</td> <td>-</td> <td>MISO pin number</td></tr>
+%%   <tr> <td>`mosi'</td> <td>`non_neg_integer()'</td> <td>-</td> <td>MOSI pin number</td></tr>
+%%   <tr> <td>`sclk'</td> <td>`non_neg_integer()'</td> <td>-</td> <td>SCLK pin number</td></tr>
+%%   <tr> <td>`peripheral'</td> <td>`hspi | vspi'</td> <td>`hspi'</td> <td>SPI Peripheral (ESP32 only)</td></tr>
 %% </table>
 %%
 %% Each device configuration is a properties list containing the following entries:
@@ -105,7 +107,7 @@
 %%   <tr> <th>Key</th> <th>Type</th> <th>Default</th> <th>Description</th> </tr>
 %%   <tr> <td>`clock_speed_hz'</td> <td>`non_neg_integer()'</td> <td>`1000000'</td> <td>Clock speed for the SPI device (in hz)</td></tr>
 %%   <tr> <td>`mode'</td> <td>`0..3'</td> <td>`0'</td> <td>SPI device mode</td></tr>
-%%   <tr> <td>`spi_cs_io_num'</td> <td>`non_neg_integer()'</td> <td>-</td> <td>SPI Chip Select pin number</td></tr>
+%%   <tr> <td>`cs'</td> <td>`non_neg_integer()'</td> <td>-</td> <td>SPI Chip Select pin number</td></tr>
 %%   <tr> <td>`address_len_bits'</td> <td>`non_neg_integer()'</td> <td>8</td> <td>Number of bits in the device address</td></tr>
 %% </table>
 %%
@@ -113,16 +115,16 @@
 %% <pre>
 %% Params = [
 %%    {bus_config, [
-%%        {miso_io_num, 16},
-%%        {mosi_io_num, 17},
-%%        {sclk_io_num, 5}
+%%        {miso, 16},
+%%        {mosi, 17},
+%%        {sclk, 5}
 %%    },
 %%    {device_config, [
 %%        {device1, [
-%%            {spi_cs_io_num, 18}
+%%            {cs, 18}
 %%        ]},
 %%        {device2, [
-%%            {spi_cs_io_num, 19}
+%%            {cs, 19}
 %%        ]}
 %%    ]}
 %% ]
@@ -298,17 +300,68 @@ validate_params(Params) ->
     throw({error, {not_a_map_or_list, Params}}).
 
 %% @private
-validate_bus_config(BusConfig) when is_map(BusConfig) orelse is_list(BusConfig) ->
+validate_bus_config(MaybeOldBusConfig) when
+    is_map(MaybeOldBusConfig) orelse is_list(MaybeOldBusConfig)
+->
+    BusConfig = migrate_deprecated(MaybeOldBusConfig),
     #{
-        miso_io_num => validate_integer_entry(miso_io_num, BusConfig),
-        mosi_io_num => validate_integer_entry(mosi_io_num, BusConfig),
-        sclk_io_num => validate_integer_entry(sclk_io_num, BusConfig),
-        spi_peripheral => validate_spi_peripheral(get_value(spi_peripheral, BusConfig, hspi))
+        miso => validate_integer_entry(miso, BusConfig),
+        mosi => validate_integer_entry(mosi, BusConfig),
+        sclk => validate_integer_entry(sclk, BusConfig),
+        peripheral => validate_peripheral(get_value(peripheral, BusConfig, hspi))
     };
 validate_bus_config(undefined) ->
     throw({badarg, missing_bus_config});
 validate_bus_config(BusConfig) ->
     throw({badarg, {not_a_map_or_list, BusConfig}}).
+
+%% @private
+migrate_deprecated(MaybeDeprecated) when is_map(MaybeDeprecated) ->
+    Iter = maps:iterator(MaybeDeprecated),
+    migrate_deprecated_iter(maps:next(Iter), #{});
+migrate_deprecated(MaybeDeprecated) when is_list(MaybeDeprecated) ->
+    lists:foldl(
+        fun migrate_deprecated_fold/2,
+        [],
+        MaybeDeprecated
+    );
+migrate_deprecated(MaybeDeprecated) ->
+    throw({bardarg, {not_a_map_or_list, MaybeDeprecated}}).
+
+%% @private
+migrate_deprecated_iter(none, Accum) ->
+    Accum;
+migrate_deprecated_iter({K, V, Iter}, Accum) ->
+    {Status, NewK} = replace_key(K),
+    warn_deprecated(Status, K, NewK),
+    migrate_deprecated_iter(maps:next(Iter), Accum#{NewK => V}).
+
+%% @private
+migrate_deprecated_fold({K, V}, Accum) ->
+    {Status, NewK} = replace_key(K),
+    warn_deprecated(Status, K, NewK),
+    [{NewK, V} | Accum];
+migrate_deprecated_fold(E, Accum) ->
+    [E | Accum].
+
+%% @private
+replace_key(Key) ->
+    case Key of
+        pico -> {rename, mosi};
+        poci -> {rename, miso};
+        miso_io_num -> {warning, miso};
+        mosi_io_num -> {warning, mosi};
+        sclk_io_num -> {warning, sclk};
+        spi_cs_io_num -> {warning, cs};
+        spi_clock_hz -> {warning, clock_speed_hz};
+        spi_peripheral -> {warning, peripheral};
+        Any -> {ok, Any}
+    end.
+
+warn_deprecated(warning, OldKey, NewKey) ->
+    io:format("SPI: found deprecated ~p, use ~p instead!!!~n", [OldKey, NewKey]);
+warn_deprecated(_Status, Key, Key) ->
+    ok.
 
 %% @private
 validate_integer_entry(Key, Map) ->
@@ -325,9 +378,9 @@ validate_is_integer(Key, Value) ->
     throw({badarg, {not_an_integer_value, {Key, Value}}}).
 
 %% @private
-validate_spi_peripheral(hspi) -> hspi;
-validate_spi_peripheral(vspi) -> vspi;
-validate_spi_peripheral(Value) -> throw({bardarg, {spi_peripheral, Value}}).
+validate_peripheral(hspi) -> hspi;
+validate_peripheral(vspi) -> vspi;
+validate_peripheral(Value) -> throw({bardarg, {peripheral, Value}}).
 
 %% @private
 validate_device_config(DeviceConfig) when is_map(DeviceConfig) ->
@@ -360,11 +413,14 @@ validate_device_config_fold(E, _Accum) ->
     throw({bardarg, {not_proplist_entry, E}}).
 
 %% @private
-validate_device_config_entries(Entries) when is_map(Entries) orelse is_list(Entries) ->
+validate_device_config_entries(MaybeOldEntries) when
+    is_map(MaybeOldEntries) orelse is_list(MaybeOldEntries)
+->
+    Entries = migrate_deprecated(MaybeOldEntries),
     #{
-        spi_clock_hz => validate_integer_entry(spi_clock_hz, Entries),
+        clock_speed_hz => validate_integer_entry(clock_speed_hz, Entries),
         mode => validate_mode(get_value(mode, Entries, undefined)),
-        spi_cs_io_num => validate_integer_entry(spi_cs_io_num, Entries, undefined),
+        cs => validate_integer_entry(cs, Entries, undefined),
         address_len_bits => validate_address_len_bits(
             get_value(address_len_bits, Entries, undefined)
         ),